change action space in game env to float32 add random speed for training phase add some wrappers add new env with distance to center env_withvisionguide

+++ l2r/envs/env.py
@@ -215,7 +215,7 @@ class RacingEnv(gym.Env):
         )
 
         # openAI gym compliance - action space
-        self.action_space = Box(low=-1.0, high=1.0, shape=(2,), dtype=np.float64)
+        self.action_space = Box(low=-1.0, high=1.0, shape=(2,), dtype=np.float32)
         self.multimodal = env_kwargs["multimodal"]
 
         # misc
